прости, но у нас нет столько времени

+++ modules/task_3/vlasov_m_shell_sort_batcher_merge/shell_sort_batcher_merge.cpp
@@ -130,7 +130,7 @@ namespace BatcherMerge {
             if (rank == comp.first) {
                 MPI_Send(part.data(), part_size, MPI_INT, comp.second, 0, MPI_COMM_WORLD);
                 MPI_Recv(part_curr.data(), part_size, MPI_INT, comp.second, 0, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
-                for (size_t i = 0, i_curr = 0, i_temp = 0; i_temp < part_size; i_temp++) {
+                for (int i = 0, i_curr = 0, i_temp = 0; i_temp < part_size; i_temp++) {
                     int value = part[i];
                     int value_curr = part_curr[i_curr];
                     if (value < value_curr) {
@@ -146,7 +146,7 @@ namespace BatcherMerge {
                 MPI_Recv(part_curr.data(), part_size, MPI_INT, comp.first, 0, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
                 MPI_Send(part.data(), part_size, MPI_INT, comp.first, 0, MPI_COMM_WORLD);
                 size_t i_start = part_size - 1;
-                for (size_t i = i_start, i_curr = i_start, i_temp = part_size; i_temp > 0; i_temp--) {
+                for (int i = i_start, i_curr = i_start, i_temp = part_size; i_temp > 0; i_temp--) {
                     int value = part[i];
                     int value_curr = part_curr[i_curr];
                     if (value > value_curr) {
